+++ classes/api.php
@@ -218,6 +218,7 @@ class api {
             'CURLOPT_RETURNTRANSFER' => true,
             'CURLOPT_TIMEOUT' => constants::REQUEST_API_TIMEOUT,
             'CURLOPT_MAXREDIRS' => 3,
+            'CURLOPT_SSL_VERIFYPEER' => true,
         ];
 
         $url = self::get_api_url() . '/' . ltrim($endpoint, '/');
@@ -484,7 +485,7 @@ class api {
             $s3url = $result['multiplarturl'] ?? null;
 
             // Make sure the returned URL is in fact an AWS S3 pre-signed URL, and we send the encryption key only to AWS.
-            if ($encryptionkey && !preg_match('|^https://[^/]+\\.s3\\.amazonaws\\.com/|', $s3url)) {
+            if ($encryptionkey && !self::is_s3_url($s3url)) {
                 throw new \moodle_exception('error_invaliduploadlink', 'tool_vault', '', $filename);
             }
 
@@ -508,7 +509,7 @@ class api {
         $uploadheaders = array_merge($encryptionheaders, $result['uploadheaders'] ?? []);
         foreach ($s3urls as $partno => $s3url) {
             // Make sure the returned URL is in fact an AWS S3 pre-signed URL, and we send the encryption key only to AWS.
-            if ($encryptionkey && !preg_match('|^https://[^/]+\\.s3\\.amazonaws\\.com/|', $s3url)) {
+            if ($encryptionkey && !self::is_s3_url($s3url)) {
                 throw new \moodle_exception('error_invaliduploadlink', 'tool_vault', '', $filename);
             }
             // Upload the file or a part of the file to the pre-signed URL.
@@ -654,7 +655,7 @@ class api {
         }
 
         // Make sure the returned URL is in fact an AWS S3 pre-signed URL, and we send the encryption key only to AWS.
-        if (!preg_match('|^https://[^/]+\\.s3\\.amazonaws\\.com/|', $s3url)) {
+        if (!self::is_s3_url($s3url)) {
             throw new \moodle_exception('error_notavalidlink', 'tool_vault', '', s($s3url));
         }
 
@@ -699,7 +700,7 @@ class api {
         $encrypted = $result['encrypted'] ?? false;
 
         // Make sure the returned URL is in fact an AWS S3 pre-signed URL, and we send the encryption key only to AWS.
-        if ($encrypted && !preg_match('|^https://[^/]+\\.s3\\.amazonaws\\.com/|', $s3url)) {
+        if ($encrypted && !self::is_s3_url($s3url)) {
             throw new \moodle_exception(
                 'error_invaliddownloadlink',
                 'tool_vault',
@@ -758,6 +759,16 @@ class api {
      */
     public static function prepare_encryption_key(?string $passphrase): string {
         return strlen($passphrase) ? base64_encode(hash('sha256', $passphrase, true)) : '';
+    }
+
+    /**
+     * Checks that the URL is an AWS S3 URL, so that the encryption key is only sent to AWS
+     *
+     * @param string|null $url
+     * @return bool
+     */
+    public static function is_s3_url(?string $url): bool {
+        return (bool)preg_match('|^https://[a-z0-9.-]+\\.s3\\.amazonaws\\.com/|', (string)$url);
     }
 
     /**